fix warning

🤖 AI[0%] 🔧 Human Fixed[0%] 🧑 Human[100%] 👌 AI Adopted[0%]

+++ rosen/modules/render_service/core/pipeline/render_thread/rs_uni_render_processor.cpp
@@ -309,7 +309,7 @@ void RSUniRenderProcessor::CreateSolidColorLayer(RSLayerPtr layer, RSSurfaceRend
     }
     GraphicSolidColorLayerProperty solidColorLayerProperty;
     if (layer->GetZorder() > 0) {
-        solidColorLayerProperty.zOrder = layer->GetZorder() - 1;
+        solidColorLayerProperty.zOrder = static_cast<int32_t>(layer->GetZorder()) - 1;
     }
     solidColorLayerProperty.transformType = GraphicTransformType::GRAPHIC_ROTATE_NONE;
     auto layerRect = layer->GetLayerSize();
